% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weight.R
\name{dist_weight}
\alias{dist_weight}
\title{Weight function for w}
\usage{
dist_weight(X, phi, dist.type, p)
}
\arguments{
\item{X}{data matrix p by n}

\item{phi}{scale}

\item{dist.type}{c("euclidean", "maximum", "manhattan", "canberra", "binary" , "minkowski")}

\item{p}{power of distance}
}
\description{
wrap up of the stat::dist function for the weight of data
}

