% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_validate.R
\name{predict_validate}
\alias{predict_validate}
\title{Predict validation data}
\usage{
predict_validate(X, cluster, features, data_valide)
}
\arguments{
\item{X}{centralized data (p by n)}

\item{cluster}{cluster assignment for each subject}

\item{features}{used features}

\item{data_valide}{validate dataset list}
}
\description{
report Rand index and variable stability in validate dataset
}

